+++ src/gtc/dace/expansion/sdfg_builder.py
@@ -27,7 +27,8 @@ from eve import NodeVisitor
 from gtc import daceir as dcir
 from gtc.dace.expansion.tasklet_codegen import TaskletCodegen
 from gtc.dace.expansion.utils import get_dace_debuginfo
-from gtc.dace.utils import data_type_to_dace_typeclass, make_dace_subset
+from gtc.dace.helpers import make_dace_subset
+from gtc.dace.symbol_utils import data_type_to_dace_typeclass
 
 
 class StencilComputationSDFGBuilder(NodeVisitor):
